Why do you have to solve the problem of The Federal Reserve System of USA ?
Why do you -as a person- you have to solve the present conjunctural problem of The Federal Reserve System of US ?
Because without persons the money is nothing. Then :
- there are no needs in monetary policy ;
- there are no needs in political production of money regulations.

Why you ?
Because things monetary, financial, spiritual nutrition and of institucional feeds, are coming from your person or are from your creation.

So because it is your legitimacy it's your responsibility to move.